partition du projet en différents modules
