+++ utilities.py
@@ -1,0 +1,14 @@
+import random
+from math import floor
+
+# Utilities
+def round_to(n, div):
+    return floor(n / div) * div
+
+
+def clamp(value, min_value, max_value):
+    return min(max_value, max(value, min_value))
+
+
+def generate_random_color():
+    return random.randrange(255), random.randrange(255), random.randrange(255)
